Time: 0 ms (100%), Space: 9.4 MB (76%) - LeetHub

+++ 0453-minimum-moves-to-equal-array-elements/0453-minimum-moves-to-equal-array-elements.c
@@ -7,6 +7,6 @@ int minMoves(int* a, int n) {
         if(a[i]<m)
             m=a[i];
     }
-    return (int)(s-m*n);
+    return s-m*n;
     
 }
